+++ packages/next-workflow-builder/src/client/components/workflow/utils/code-generators.ts
@@ -6,12 +6,14 @@ import { findActionById } from "../../../../plugins";
 import { conditionAction } from "../../../../plugins/condition";
 import { databaseQueryAction } from "../../../../plugins/database-query";
 import { httpRequestAction } from "../../../../plugins/http-request";
+import { switchAction } from "../../../../plugins/switch";
 
 // System action templates (non-plugin actions)
 const SYSTEM_ACTION_TEMPLATES: Record<string, string> = {
   "Database Query": databaseQueryAction.codeGenerator,
   "HTTP Request": httpRequestAction.codeGenerator,
   Condition: conditionAction.codeGenerator,
+  Switch: switchAction.codeGenerator,
 };
 
 const FALLBACK_ACTION_CODE = `async function actionStep(input: Record<string, unknown>) {